Remove gevent monkey patching (#14)
In the past I added it to solve issues with the Steam library running in
a different thread. I am not sure why I do not need it anymore, but it
works perfectly without it now. Not using monkey patching also solves
some massive threading stacktraces when terminating the bot.